move auth api to feature (auth page refactor coming...)
