+++ src/features/auth/api/auth.api.ts
@@ -1,6 +1,6 @@
-import { currentUser, setUser } from "@/features/auth/model/auth.store"
+import { setUser } from "@/features/auth/model/auth.store"
 import { API } from "@/shared/api"
-import type { User } from "@/shared/models/user"
+import type { User } from "@/entities/user"
 import type { UUIDv4 } from "@/shared/lib/uuid"
 
 type AuthCredentials = {
@@ -21,7 +21,6 @@ type LoginData = {
 
 export async function signIn(data: AuthCredentials) {
     let loginData: LoginData = await API.post('auth/login', data);
-    API.defaultHeaders["Authorization"] = `Bearer: ${loginData.token}`;
     localStorage.setItem('token', loginData.token);
     let user: User = await API.get(`users/${loginData.user_id}`);
     setUser(user);
@@ -30,7 +29,7 @@ export async function signIn(data: AuthCredentials) {
 
 export async function signUp(data: SignUpData) {
     let success = await API.post('auth/signup', data);
-    if(success == "Success") console.log("Sign up successful");
+    if (success == "Success") console.log("Sign up successful");
     else return false;
     return true;
 }
